# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a Catopia API of CPE Senior Project.",
        "title": "Catopia API",
        "termsOfService": "http://somewhere.com/",
        "contact": {
            "name": "CPE34 - Catopia"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/forget-password": {
            "post": {
                "description": "Forget password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forgot Password"
                ],
                "operationId": "UserForgetPasswordHandler",
                "parameters": [
                    {
                        "description": "username",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.GetOTP"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
        "/api/reset-password": {
            "put": {
                "description": "Reset password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Forgot Password"
                ],
                "operationId": "UserResetPasswordHandler",
                "parameters": [
                    {
                        "description": "code and password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdatePassword"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/treatment/:user_id/{cat_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get treatment by cat ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentGetByCatIDHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cat id",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Treatment"
                            }
                        }
                    }
                }
            }
        },
        "/api/treatment/type": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get treatment type",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentGetTypeHandler",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.TreatmentType"
                        }
                    }
                }
            }
        },
        "/api/treatment/{cat_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create new treatment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentCreateHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cat id",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "create treatment",
                        "name": "createTreatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.CreateTreatment"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/treatment/{cat_id}/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get treatment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentGetByIDHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cat id",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of treatment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Treatment"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update treatment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentUpdateHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cat id",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of treatment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update treatment",
                        "name": "updateTreatment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdateTreatment"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete treatment by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "treatment"
                ],
                "operationId": "TreatmentDeleteHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "cat id",
                        "name": "cat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id of treatment",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "UserGetHandler",
                "responses": {}
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update user detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "UserUpdateHandler",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "user data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UpdateUser"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/api/user/answer": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "User answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "operationId": "UserAnswerHandler",
                "parameters": [
                    {
                        "description": "user answer",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UserAnswer"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "domain.Treatment": {
            "type": "object",
            "properties": {
                "catID": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "treatmentTypeID": {
                    "type": "integer"
                },
                "vet": {
                    "type": "string"
                }
            }
        },
        "domain.TreatmentType": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "treatment_type": {
                    "type": "string"
                }
            }
        },
...
        "payload.CreateTreatment": {
            "type": "object",
            "required": [
                "date",
                "treatmentTypeID"
            ],
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "detail": {
                    "type": "string",
                    "example": "vaccination"
                },
                "location": {
                    "type": "string",
                    "example": "clinic"
                },
                "treatmentTypeID": {
                    "type": "integer",
                    "example": 1
                },
                "vet": {
                    "type": "string",
                    "example": "Dr. John Doe"
                }
            }
        },
        "payload.GetOTP": {
            "type": "object",
            "required": [
                "username"
            ],
            "properties": {
                "username": {
                    "type": "string",
                    "example": "kongsakchai"
                }
            }
        },
...
        "payload.UpdatePassword": {
            "type": "object",
            "required": [
                "code",
                "password"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123456"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "payload.UpdateTreatment": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "detail": {
                    "type": "string",
                    "example": "vaccination"
                },
                "location": {
                    "type": "string",
                    "example": "clinic"
                },
                "treatmentTypeID": {
                    "type": "integer",
                    "example": 1
                },
                "vet": {
                    "type": "string",
                    "example": "Dr. John Doe"
                }
            }
        },
        "payload.UpdateUser": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string",
                    "format": "date",
                    "example": "2021-01-20"
                },
                "email": {
                    "type": "string",
                    "example": "mail@mail.com"
                },
                "gender": {
                    "type": "string",
                    "enum": [
                        "male",
                        "female"
                    ],
                    "example": "male"
                },
                "password": {
                    "type": "string",
                    "example": "password123"
                },
                "profile": {
                    "type": "string",
                    "example": "url of image"
                },
                "username": {
                    "type": "string",
                    "example": "kongsakchai"
                }
            }
        },
        "payload.UserAnswer": {
            "type": "object",
            "required": [
                "answer"
            ],
            "properties": {
                "answer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "1",
                        "2",
                        "3",
                        "4",
                        "5"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}